+++ src/cli/commands/config/init.rs
@@ -1,3 +1,5 @@
+// Copyright 2022-2023 Protocol Labs
+// SPDX-License-Identifier: MIT
 use crate::cli::{CommandLineHandler, GlobalArguments};
 use crate::config::DEFAULT_CONFIG_TEMPLATE;
 use async_trait::async_trait;
