+++ src/cli/commands/wallet/new.rs
